Use shell-script for kexec-wrap
The C code mainly called busybox functions, so we can just
use a shell script instead. That means we save a few lines
of code and do not need an Android cross-compiler toolchain.

As a side-effect the output to /dev/ttyO2 by the script
seems to be more stable than the one of the C code.